organize src fils by pages in frontend
