+++ hp-mri-frontend/src/assets/analyze_image.png
