+++ hp-mri-frontend/src/assets/image_analysis.png
