+++ hp-mri-frontend/src/assets/medcap.png
