+++ hp-mri-frontend/src/assets/medcap_top_image.png
