+++ hp-mri-frontend/src/assets/metabolic_model.png
