+++ hp-mri-frontend/src/assets/mr_calc_image.png
