+++ hp-mri-frontend/src/assets/mri_hardware.png
