+++ hp-mri-frontend/src/assets/mri_recon.png
