+++ hp-mri-frontend/src/assets/mri_simulator.png
