+++ hp-mri-frontend/src/assets/penn_building_image.png
