+++ hp-mri-frontend/src/assets/pigi_logo.png
