+++ hp-mri-frontend/src/assets/simulate_image.png
