+++ hp-mri-frontend/src/components/HeaderAccount.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { FaUserCircle } from 'react-icons/fa';
 import { AppBar, Toolbar, Typography, IconButton, Box, Avatar, Tooltip, useTheme } from '@mui/material';
-import PigiLogo from './../images/pigi_optblue_transparentexceptpennlogo.png';
+import PigiLogo from './../assets/pigi_optblue_transparentexceptpennlogo.png';
 // import TheMedcap from './../images/the-medcap.png';
-import Medcap from './../images/medcap.png'
+import Medcap from './../assets/medcap.png'
 
 const HeaderAccount: React.FC = () => {
   const theme = useTheme();
