+++ hp-mri-frontend/src/pages/mrdpages/MRDFileDetails.tsx
@@ -45,9 +45,7 @@ const MRDFileDetails: React.FC = () => {
         }
 
         axios
-            // localhost API call
-            // .get(`http://127.0.0.1:5000/api/mrd-files/${fileId}`)
-            .get('/api/mrd-files/${fileId}')
+            .get(`/api/mrd-files/${fileId}`)
             .then((response) => {
                 setFileDetails(response.data);
                 setEditedTags({ parameter: response.data.parameter, raw: response.data.raw?.description });
@@ -58,7 +56,6 @@ const MRDFileDetails: React.FC = () => {
 
     const fetchImages = () => {
         axios
-            // .get(`http://127.0.0.1:5000/api/images/${fileId}`)
             .get(`/api/images/${fileId}`) 
             .then((response) => setImages(response.data))
             .catch((error) => console.error('Error fetching images:', error));
